requirements/base.txt
# https://github.com/django/django/archive/1.7c1.tar.gz
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
South==0.8.4
MySQL-Python
